poc/serve/requirements.txt
google-genai==0.3.0
chromadb==0.4.24
python-dotenv==1.0.1
numpy<2.0.0
